4.21

+++ part4/bloglist/controllers/blogs.js
@@ -10,8 +10,8 @@ blogsRouter.get('/', async (request, response, next) => {
       .populate('user', { username: 1, name: 1 });
     response.json(blogs.map(blog => blog.toJSON()));
   }
-  catch(error) {
-    next(error);
+  catch(exception) {
+    next(exception);
   }
 });
 
@@ -39,8 +39,8 @@ blogsRouter.post('/', async (request, response, next) => {
     await user.save();
     response.json(savedBlog.toJSON());
   }
-  catch(error) {
-    next(error);
+  catch(exception) {
+    next(exception);
   }
 });
 
@@ -57,18 +57,27 @@ blogsRouter.put('/:id', async (request, response, next) => {
     const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true });
     response.json(updatedBlog.toJSON());
   }
-  catch (error) {
-    next(error);
+  catch (exception) {
+    next(exception);
   }
 });
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!decodedToken) {
+      return response.status(401).json({ error: 'token missing or invalid' });
+    }
+
+    const blog = await Blog.findById(request.params.id);
+    if (blog.user.toString() !== decodedToken.id)
+      return response.status(403).json({ error: 'blogs can only be deleted by the user who added them' });
+
     await Blog.findByIdAndRemove(request.params.id);
     response.status(204).end();
   }
-  catch (error) {
-    next(error);
+  catch (exception) {
+    next(exception);
   }
 });
